requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
scikit-learn==1.3.2
pandas==2.2.2
nltk==3.9.0
joblib==1.3.2
textblob==0.17.1
python-dotenv==1.0.0
